docs/examples/multi-agent/requirements.txt
# Multi-Agent A2A Communication Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0